Refactor AddPlant component
Refactored the `AddPlant` component into smaller, more manageable files.  Removed any unused imports or files.
